Export genes plot component
