+++ packages/track-genes/src/index.js
@@ -1,1 +1,2 @@
+export { GenesPlot } from './GenesPlot'
 export { GenesTrack } from './GenesTrack'
